refactor(generate-docs): add page key constants; extract service loader
- Define named constants for page keys in docsconfig.go, replacing
magic string literals across all generator files
- Move service loading types and functions into a dedicated
serviceloader.go, leaving services.go focused on generation only
- Fix loadService to parse YAML once via node decode instead of two
separate Unmarshal calls
- Inline single-use writeConfigSections helper into its call site
- Use strings.Builder in generateConfigSections to avoid slice joins
