+++ cmd/generate-docs/contributing.go
@@ -11,9 +11,9 @@ func generateContributingGuide() error {
 		return fmt.Errorf("read CONTRIBUTING.md: %w", err)
 	}
 
-	out, err := formatDocument(pageFM("contributing"), string(data))
+	out, err := formatDocument(pageFM(pageContributing), string(data))
 	if err != nil {
 		return err
 	}
-	return writeOutput(pageOutput("contributing"), out)
+	return writeOutput(pageOutput(pageContributing), out)
 }
