+++ cmd/generate-docs/services.go
@@ -2,114 +2,9 @@ package main
 
 import (
 	"fmt"
-	"os"
-	"path/filepath"
 	"sort"
 	"strings"
-
-	"gopkg.in/yaml.v3"
 )
-
-type serviceConfig struct {
-	Name               string                `yaml:"name"`
-	Description        string                `yaml:"description"`
-	Hidden             bool                  `yaml:"hidden"`
-	Environment        map[string]string     `yaml:"environment"`
-	Documentation      *serviceDocumentation `yaml:"documentation"`
-	configSchemaFields []*schemaField
-}
-
-type serviceDocumentation struct {
-	UseCases []string `yaml:"use_cases"`
-	Examples []string `yaml:"examples"`
-}
-
-type loadedService struct {
-	name     string
-	config   serviceConfig
-	category string
-}
-
-type categoryConfig struct {
-	Icon  string `yaml:"icon"`
-	Order int    `yaml:"order"`
-}
-
-func getCategoryConfig(name string) categoryConfig {
-	if c, ok := docs.Categories[name]; ok {
-		return c
-	}
-	return docs.Categories["other"]
-}
-
-func loadAllServices() ([]loadedService, error) {
-	var services []loadedService
-	err := filepath.Walk(servicesDirPath, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-		if info.IsDir() {
-			return nil
-		}
-		ext := filepath.Ext(path)
-		if ext != ".yaml" && ext != ".yml" {
-			return nil
-		}
-		svc, err := loadService(path)
-		if err != nil {
-			return err
-		}
-		if svc != nil {
-			services = append(services, *svc)
-		}
-		return nil
-	})
-	return services, err
-}
-
-func loadService(path string) (*loadedService, error) {
-	data, err := os.ReadFile(path)
-	if err != nil {
-		return nil, fmt.Errorf("read %s: %w", path, err)
-	}
-
-	var svc serviceConfig
-	if err := yaml.Unmarshal(data, &svc); err != nil {
-		return nil, fmt.Errorf("parse %s: %w", path, err)
-	}
-	if svc.Hidden {
-		return nil, nil
-	}
-
-	var rootNode yaml.Node
-	if err := yaml.Unmarshal(data, &rootNode); err != nil {
-		return nil, fmt.Errorf("parse node %s: %w", path, err)
-	}
-	svc.configSchemaFields = extractSchemaFields(nodeGet(&rootNode, keyConfigSchema))
-
-	ext := filepath.Ext(path)
-	name := strings.TrimSuffix(filepath.Base(path), ext)
-	return &loadedService{name: name, config: svc, category: inferCategory(path)}, nil
-}
-
-func inferCategory(path string) string {
-	relPath, _ := filepath.Rel(servicesDirPath, path)
-	parts := strings.Split(filepath.ToSlash(relPath), "/")
-	if len(parts) >= 2 {
-		return parts[0]
-	}
-	return "other"
-}
-
-func indexServices(services []loadedService) map[string]loadedService {
-	svcMap := make(map[string]loadedService, len(services))
-	for _, svc := range services {
-		svcMap[svc.name] = svc
-	}
-	return svcMap
-}
-
-// ---- Generator: services-guide ----
 
 func generateServicesGuide() error {
 	services, err := loadAllServices()
@@ -120,7 +15,7 @@ func generateServicesGuide() error {
 	byCategory := groupByCategory(services)
 	categories := sortedCategories(byCategory)
 
-	page := docs.Pages["services"]
+	page := docs.Pages[pageServices]
 	var sb strings.Builder
 	sb.WriteString("# " + page.Heading + "\n\n")
 	sb.WriteString(fmt.Sprintf(page.ServiceCount, len(services)) + "\n\n")
@@ -138,30 +33,11 @@ func generateServicesGuide() error {
 		}
 	}
 
-	out, err := formatDocument(pageFM("services"), sb.String())
+	out, err := formatDocument(pageFM(pageServices), sb.String())
 	if err != nil {
 		return err
 	}
-	return writeOutput(pageOutput("services"), out)
-}
-
-func groupByCategory(services []loadedService) map[string][]loadedService {
-	byCategory := make(map[string][]loadedService)
-	for _, svc := range services {
-		byCategory[svc.category] = append(byCategory[svc.category], svc)
-	}
-	return byCategory
-}
-
-func sortedCategories(byCategory map[string][]loadedService) []string {
-	categories := make([]string, 0, len(byCategory))
-	for cat := range byCategory {
-		categories = append(categories, cat)
-	}
-	sort.Slice(categories, func(i, j int) bool {
-		return getCategoryConfig(categories[i]).Order < getCategoryConfig(categories[j]).Order
-	})
-	return categories
+	return writeOutput(pageOutput(pageServices), out)
 }
 
 func renderServiceSection(svc loadedService) string {
